refactor(profiling-analysis): split common.py, YAML-drive rules, unify html data layer
- common.py 2191 -> facade (232) + models/store/sources/pipeline/work/
  rules/metrics modules; duplicated helpers consolidated
- categories_and_roles now driven by kernel_signatures.yaml match_rules
  (11909-case differential corpus, 0 mismatch); resolve_attention_family
  driven by attention_families.yaml resolver; parity tests added
- html_report reuses the shared data layer instead of a second Event/Bundle
- diagnosis thresholds/metadata externalized to diagnosis_rules.yaml
  (A/B byte-identical findings)
- drop consumer-less schemas/analysis_bundle.schema.json; docs updated
Tests: 271 passed.
