+++ .agents/skills/ascend-profiling-analysis/scripts/ascend_profile/diagnostics.py
@@ -1,41 +1,111 @@
 #!/usr/bin/env python3
-"""Generate diagnosis claims from summary and cross-rank evidence tables."""
+"""Generate diagnosis claims from summary and cross-rank evidence tables.
+
+Thresholds, severity/confidence labels, message templates, and static
+limitations live in ``knowledge/diagnosis_rules.yaml`` (loaded at runtime by
+``load_diagnosis_rules``). The *condition logic* — which rows trigger which
+finding — stays here in Python on purpose; the YAML carries constants and
+wording only, no rule DSL.
+"""
 
 from __future__ import annotations
 
 import argparse
-import json
+import functools
 from collections import defaultdict
 from pathlib import Path
 from typing import Any, Mapping, Sequence
 
 try:
     from .common import DiagnosisFinding, SCHEMA_VERSION, TOOL_VERSION, csv_rows, emit_stage_json, stable_id, utc_now, write_json
+    from .store import KNOWLEDGE_DIR, parse_jsonish
 except ImportError:  # pragma: no cover
     import sys
 
     sys.path.insert(0, str(Path(__file__).resolve().parent))
     from common import DiagnosisFinding, SCHEMA_VERSION, TOOL_VERSION, csv_rows, emit_stage_json, stable_id, utc_now, write_json  # type: ignore[no-redef]
+    from store import KNOWLEDGE_DIR, parse_jsonish  # type: ignore[no-redef]
 
 
-CROSS_RANK_SKEW_RATIO = 2.0
-HIGH_SKEW_RATIO = 4.0
-CROSS_RANK_SKEW_US = 1000.0
-DP_WALL_SKEW_RATIO = 2.0
+DIAGNOSIS_RULES_PATH = KNOWLEDGE_DIR / "diagnosis_rules.yaml"
+
+_FINDING_KEYS = frozenset({
+    "scope",
+    "severity",
+    "escalated_severity",
+    "confidence",
+    "summary",
+    "summary_template",
+    "limitations",
+    "limitations_if_no_evidence",
+    "limitation_parts",
+})
 
 
-def parse_jsonish(value: Any, default: Any) -> Any:
-    if value is None:
-        return default
-    if not isinstance(value, str):
-        return value
-    text = value.strip()
-    if not text:
-        return default
+def load_diagnosis_rules(path: Path = DIAGNOSIS_RULES_PATH) -> dict[str, Any]:
+    """Load and validate ``knowledge/diagnosis_rules.yaml``.
+
+    This is the single source of truth for finding thresholds and wording,
+    not a silent fallback: if the file is missing or malformed we raise
+    instead of guessing.
+    """
+
     try:
-        return json.loads(text)
-    except json.JSONDecodeError:
-        return default
+        import yaml  # type: ignore[import-untyped]
+    except ImportError as exc:  # pragma: no cover - yaml is a hard dependency
+        raise RuntimeError("PyYAML is required to load knowledge/diagnosis_rules.yaml") from exc
+    if not path.exists():
+        raise RuntimeError(f"diagnosis rules knowledge base missing: {path}")
+    doc = yaml.safe_load(path.read_text(encoding="utf-8")) or {}
+
+    thresholds = doc.get("thresholds")
+    if not isinstance(thresholds, dict) or not thresholds:
+        raise RuntimeError(f"{path.name}: thresholds: missing or not a mapping")
+    for key, value in thresholds.items():
+        if not isinstance(value, (int, float)) or isinstance(value, bool):
+            raise RuntimeError(f"{path.name}: thresholds.{key}: must be numeric, got {value!r}")
+
+    findings = doc.get("findings")
+    if not isinstance(findings, dict) or not findings:
+        raise RuntimeError(f"{path.name}: findings: missing or not a mapping")
+    for finding_type, meta in findings.items():
+        context = f"{path.name}: findings.{finding_type}"
+        if not isinstance(meta, dict):
+            raise RuntimeError(f"{context}: must be a mapping")
+        unknown = set(meta) - _FINDING_KEYS
+        if unknown:
+            raise RuntimeError(f"{context}: unknown keys: {sorted(unknown)}")
+        for key in ("scope", "severity", "confidence"):
+            if not isinstance(meta.get(key), str) or not meta[key]:
+                raise RuntimeError(f"{context}: {key} must be a non-empty string")
+        if ("summary" in meta) == ("summary_template" in meta):
+            raise RuntimeError(f"{context}: exactly one of summary / summary_template is required")
+        if "escalated_severity" in meta and not isinstance(meta["escalated_severity"], str):
+            raise RuntimeError(f"{context}: escalated_severity must be a string")
+        for key in ("limitations", "limitations_if_no_evidence"):
+            value = meta.get(key)
+            if value is not None and (not isinstance(value, list) or not all(isinstance(item, str) for item in value)):
+                raise RuntimeError(f"{context}: {key} must be a list of strings")
+        parts = meta.get("limitation_parts")
+        if parts is not None and (not isinstance(parts, dict) or not all(isinstance(v, str) for v in parts.values())):
+            raise RuntimeError(f"{context}: limitation_parts must be a mapping of strings")
+    return {"thresholds": dict(thresholds), "findings": dict(findings)}
+
+
+@functools.lru_cache(maxsize=1)
+def _rules() -> dict[str, Any]:
+    return load_diagnosis_rules()
+
+
+def _thresholds() -> dict[str, float]:
+    return _rules()["thresholds"]
+
+
+def _meta(finding_type: str) -> Mapping[str, Any]:
+    meta = _rules()["findings"].get(finding_type)
+    if meta is None:  # pragma: no cover - guards Python/YAML drift
+        raise RuntimeError(f"diagnosis_rules.yaml: no findings.{finding_type} entry")
+    return meta
 
 
 def as_float(row: Mapping[str, Any], key: str, default: float = 0.0) -> float:
@@ -75,7 +145,14 @@ def finding(
     )
 
 
+def _skew_severity(meta: Mapping[str, Any], duration_ratio: float) -> str:
+    if duration_ratio >= _thresholds()["high_skew_ratio"]:
+        return str(meta.get("escalated_severity") or meta["severity"])
+    return str(meta["severity"])
+
+
 def diagnose_cross_rank(alignment_rows: Sequence[Mapping[str, Any]]) -> list[DiagnosisFinding]:
+    thresholds = _thresholds()
     findings: list[DiagnosisFinding] = []
     for row in alignment_rows:
         alignment_id = str(row.get("alignment_id") or "")
@@ -87,54 +164,58 @@ def diagnose_cross_rank(alignment_rows: Sequence[Mapping[str, Any]]) -> list[Dia
         start_skew = as_float(row, "start_skew_us")
         is_structure_mismatch = str(row.get("is_structure_mismatch")).lower() == "true"
         if alignment_type == "time_window" and is_structure_mismatch:
+            meta = _meta("rank_workload_asymmetry")
             findings.append(
                 finding(
                     finding_type="rank_workload_asymmetry",
-                    scope="cross_rank",
-                    summary="Aligned ranks show different step family or layer-count structure in the same time window.",
-                    severity="medium",
-                    confidence="medium",
+                    scope=str(meta["scope"]),
+                    summary=str(meta["summary"]),
+                    severity=str(meta["severity"]),
+                    confidence=str(meta["confidence"]),
                     rank_ids=rank_ids,
                     alignment_ids=(alignment_id,),
                     metrics=dict(row),
                 )
             )
-        if role == "communication.collective" and (duration_ratio >= CROSS_RANK_SKEW_RATIO or duration_skew >= CROSS_RANK_SKEW_US):
+        if role == "communication.collective" and (duration_ratio >= thresholds["cross_rank_skew_ratio"] or duration_skew >= thresholds["cross_rank_skew_us"]):
+            meta = _meta("communication_collective_slow")
             findings.append(
                 finding(
                     finding_type="communication_collective_slow",
-                    scope="cross_rank",
-                    summary="Aligned collective communication shows large duration skew across ranks.",
-                    severity="high" if duration_ratio >= HIGH_SKEW_RATIO else "medium",
-                    confidence="medium",
+                    scope=str(meta["scope"]),
+                    summary=str(meta["summary"]),
+                    severity=_skew_severity(meta, duration_ratio),
+                    confidence=str(meta["confidence"]),
                     rank_ids=rank_ids,
                     alignment_ids=(alignment_id,),
                     metrics=dict(row),
                 )
             )
         if role in {"moe.dispatch_expert_compute", "moe.dispatch_or_combine"} and (
-            duration_ratio >= CROSS_RANK_SKEW_RATIO or duration_skew >= CROSS_RANK_SKEW_US
+            duration_ratio >= thresholds["cross_rank_skew_ratio"] or duration_skew >= thresholds["cross_rank_skew_us"]
         ):
+            meta = _meta("ep_load_imbalance_suspected")
             findings.append(
                 finding(
                     finding_type="ep_load_imbalance_suspected",
-                    scope="cross_rank",
-                    summary="Aligned MoE dispatch/combine work shows large duration skew across ranks.",
-                    severity="high" if duration_ratio >= HIGH_SKEW_RATIO else "medium",
-                    confidence="medium",
+                    scope=str(meta["scope"]),
+                    summary=str(meta["summary"]),
+                    severity=_skew_severity(meta, duration_ratio),
+                    confidence=str(meta["confidence"]),
                     rank_ids=rank_ids,
                     alignment_ids=(alignment_id,),
                     metrics=dict(row),
                 )
             )
-        if role == "compute.matmul" and start_skew >= CROSS_RANK_SKEW_US:
+        if role == "compute.matmul" and start_skew >= thresholds["cross_rank_skew_us"]:
+            meta = _meta("slow_rank_suspected")
             findings.append(
                 finding(
                     finding_type="slow_rank_suspected",
-                    scope="cross_rank",
-                    summary="Aligned matmul work has similar operator/shape signature but large launch-time skew.",
-                    severity="medium",
-                    confidence="medium",
+                    scope=str(meta["scope"]),
+                    summary=str(meta["summary"]),
+                    severity=str(meta["severity"]),
+                    confidence=str(meta["confidence"]),
                     rank_ids=rank_ids,
                     alignment_ids=(alignment_id,),
                     metrics=dict(row),
@@ -149,44 +230,47 @@ def diagnose_rank_workload(rank_rows: Sequence[Mapping[str, Any]], step_rows: Se
     if attention_by_rank and any(attention_by_rank.values()) and not all(attention_by_rank.values()):
         reduced = [rank for rank, has_attention in attention_by_rank.items() if not has_attention]
         full = [rank for rank, has_attention in attention_by_rank.items() if has_attention]
+        meta = _meta("reduced_work_or_dummy_rank")
         findings.append(
             finding(
                 finding_type="reduced_work_or_dummy_rank",
-                scope="cross_rank",
-                summary="Some ranks lack attention/body evidence while other ranks contain full attention workload evidence.",
-                severity="medium",
-                confidence="medium",
+                scope=str(meta["scope"]),
+                summary=str(meta["summary"]),
+                severity=str(meta["severity"]),
+                confidence=str(meta["confidence"]),
                 rank_ids=tuple(sorted(reduced + full)),
                 metrics={"full_work_ranks": full, "reduced_work_candidate_ranks": reduced},
-                limitations=("This is structural evidence only; semantic dummy-run labeling requires workload context.",),
+                limitations=tuple(meta.get("limitations") or ()),
             )
         )
     wall_by_rank = {str(row.get("rank_id")): as_float(row, "wall_ms") for row in rank_rows}
     if wall_by_rank:
         values = [value for value in wall_by_rank.values() if value > 0]
-        if values and max(values) / max(1e-6, min(values)) >= DP_WALL_SKEW_RATIO:
+        if values and max(values) / max(1e-6, min(values)) >= _thresholds()["dp_wall_skew_ratio"]:
+            meta = _meta("dp_workload_imbalance")
             findings.append(
                 finding(
                     finding_type="dp_workload_imbalance",
-                    scope="cross_rank",
-                    summary="Rank-level capture wall time differs significantly; check DP workload and T-axis/shape distribution.",
-                    severity="medium",
-                    confidence="low",
+                    scope=str(meta["scope"]),
+                    summary=str(meta["summary"]),
+                    severity=str(meta["severity"]),
+                    confidence=str(meta["confidence"]),
                     rank_ids=tuple(sorted(wall_by_rank)),
                     metrics={"rank_wall_ms": wall_by_rank, "wall_ratio": max(values) / max(1e-6, min(values))},
-                    limitations=("Wall-time skew alone is not root cause evidence; shape-level corroboration is required.",),
+                    limitations=tuple(meta.get("limitations") or ()),
                 )
             )
     for row in step_rows:
         tags = parse_jsonish(row.get("anomaly_tags"), [])
         if "DEVICE_IDLE_GAP_HEAVY" in tags or "INTERNAL_BUBBLE_HEAVY" in tags:
+            meta = _meta("device_idle_bubble")
             findings.append(
                 finding(
                     finding_type="device_idle_bubble",
-                    scope="step",
-                    summary=f"Step {row.get('segment_id')} has heavy device idle bubbles.",
-                    severity="medium",
-                    confidence="high",
+                    scope=str(meta["scope"]),
+                    summary=str(meta["summary_template"]).format(segment_id=row.get("segment_id")),
+                    severity=str(meta["severity"]),
+                    confidence=str(meta["confidence"]),
                     rank_ids=(str(row.get("rank_id")),),
                     evidence_ids=tuple(parse_jsonish(row.get("evidence_ids"), [])),
                     metrics=dict(row),
@@ -203,10 +287,13 @@ def diagnose_recurring_bubbles(rank_rows: Sequence[Mapping[str, Any]], step_rows
     The flag and the dominant idle family are computed in
     ``summarize.recurring_bubble_rollup`` and read back from
     ``rank_summary.csv``; per-step evidence ids come from the bubbling
-    steps in ``step_summary.csv`` (capped at 8 so the finding stays
+    steps in ``step_summary.csv`` (capped at
+    ``thresholds.recurring_bubble_evidence_cap`` so the finding stays
     readable).
     """
 
+    evidence_cap = int(_thresholds()["recurring_bubble_evidence_cap"])
+    meta = _meta("recurring_bubble_pattern")
     bubbling_evidence_by_rank: dict[str, list[str]] = defaultdict(list)
     for row in step_rows:
         if row.get("segment_type") != "step":
@@ -222,24 +309,21 @@ def diagnose_recurring_bubbles(rank_rows: Sequence[Mapping[str, Any]], step_rows
         if str(row.get("recurring_bubble_pattern")).lower() != "true":
             continue
         rank_id = str(row.get("rank_id") or "")
-        evidence = tuple(bubbling_evidence_by_rank.get(rank_id, [])[:8])
+        evidence = tuple(bubbling_evidence_by_rank.get(rank_id, [])[:evidence_cap])
         limitations: tuple[str, ...] = ()
         if not evidence:
-            limitations = (
-                "Aggregate over step_summary.csv rows for this rank; no per-step "
-                "evidence id was available to reference.",
-            )
+            limitations = tuple(meta.get("limitations_if_no_evidence") or ())
         findings.append(
             finding(
                 finding_type="recurring_bubble_pattern",
-                scope="rank",
-                summary=(
-                    f"Rank {rank_id} shows device-idle bubbles in "
-                    f"{as_float(row, 'bubble_recurrence_ratio') * 100:.0f}% of its steps "
-                    f"(dominant idle pattern: {row.get('dominant_idle_pattern')})."
+                scope=str(meta["scope"]),
+                summary=str(meta["summary_template"]).format(
+                    rank_id=rank_id,
+                    recurrence_pct=as_float(row, "bubble_recurrence_ratio") * 100,
+                    dominant_idle_pattern=row.get("dominant_idle_pattern"),
                 ),
-                severity="medium",
-                confidence="medium",
+                severity=str(meta["severity"]),
+                confidence=str(meta["confidence"]),
                 rank_ids=(rank_id,),
                 evidence_ids=evidence,
                 metrics={
@@ -265,35 +349,22 @@ def diagnose_profile(output_dir: Path) -> dict[str, Any]:
     findings.extend(diagnose_recurring_bubbles(rank_rows, step_rows))
     for row in wait_rows:
         if str(row.get("is_false_hotspot_risk")).lower() == "true":
-            # ``wait_anchor_ops.csv`` is itself the per-kernel evidence
-            # (with ``row_ranges`` + ``sample_event_ids`` pointing back
-            # to the normalized event index). The evidence-chain
-            # validator only knows about ``evidence_index.csv``
-            # (``evd_*`` ids), so we expose the wait_anchor row's own
-            # back-references as an explicit ``limitations`` string
-            # instead of faking a cross-ID match.
+            meta = _meta("wait_anchor_false_hotspot")
+            parts = meta["limitation_parts"]
             row_ranges = str(row.get("row_ranges") or "").strip()
             sample_evt = str(row.get("sample_event_ids") or "").strip()
-            limitation_parts = [
-                "Wait-anchor false-hotspot is derived directly from "
-                "wait_anchor_ops.csv (per-kernel aggregate); not aligned "
-                "to an evidence_index.csv span.",
-            ]
+            limitation_parts = [str(parts["base"])]
             if row_ranges:
-                limitation_parts.append(
-                    f"Source rows in kernel_details.csv: {row_ranges}."
-                )
+                limitation_parts.append(str(parts["row_ranges"]).format(row_ranges=row_ranges))
             if sample_evt:
-                limitation_parts.append(
-                    f"Sample event ids in normalized_event_index.csv: {sample_evt}."
-                )
+                limitation_parts.append(str(parts["sample_event_ids"]).format(sample_event_ids=sample_evt))
             findings.append(
                 finding(
                     finding_type="wait_anchor_false_hotspot",
-                    scope="operator",
-                    summary=f"Operator {row.get('name')} has high wait ratio and low execution duration.",
-                    severity="low",
-                    confidence="high",
+                    scope=str(meta["scope"]),
+                    summary=str(meta["summary_template"]).format(name=row.get("name")),
+                    severity=str(meta["severity"]),
+                    confidence=str(meta["confidence"]),
                     rank_ids=(str(row.get("rank_id")),),
                     metrics=dict(row),
                     limitations=(" ".join(limitation_parts),),
@@ -301,25 +372,17 @@ def diagnose_profile(output_dir: Path) -> dict[str, Any]:
             )
     for row in aicpu_rows:
         if str(row.get("classification")) == "AICPU_EXPOSED_NOT_ALLOWED":
+            meta = _meta("aicpu_exposed")
             findings.append(
                 finding(
                     finding_type="aicpu_exposed",
-                    scope="operator",
-                    summary=f"AICPU operator {row.get('name')} appears exposed rather than hidden by AI Core work.",
-                    severity="medium",
-                    confidence="medium",
+                    scope=str(meta["scope"]),
+                    summary=str(meta["summary_template"]).format(name=row.get("name")),
+                    severity=str(meta["severity"]),
+                    confidence=str(meta["confidence"]),
                     rank_ids=(str(row.get("rank_id")),),
                     metrics=dict(row),
-                    # No cross-rank alignment row / evidence row is produced
-                    # for AICPU-exposure findings: the heuristic looks at one
-                    # rank at a time and uses the aicpu_summary row directly.
-                    # Surface that as an explicit limitation so the
-                    # evidence-chain validator can accept the finding.
-                    limitations=(
-                        "Derived from aicpu_summary.csv per-rank rollup; no "
-                        "cross-rank alignment_id is produced for AICPU-exposed "
-                        "findings. Trace back via the kernel name + rank_id.",
-                    ),
+                    limitations=tuple(meta.get("limitations") or ()),
                 )
             )
     payload = {
